# simplify FormatEmoji() algorithm