- [+] add Benchmarks for FormatEmoji
- [#] simplify FormatEmoji() algorithm