@suntong 2021-07-28
```
$ benchstat -delta-test ttest /tmp/new.log /tmp/org.log
name                  old time/op    new time/op    delta
FormatEmojiString-2     4.98µs ± 2%    6.73µs ± 2%     +35.10%  (p=0.000 n=3+3)
FormatEmojiString-4     5.13µs ± 2%    7.27µs ± 4%     +41.73%  (p=0.001 n=3+3)
FormatEmojiString-8     5.54µs ± 2%    7.46µs ± 2%     +34.67%  (p=0.000 n=3+3)
FormatEmojiString-12    5.80µs ± 1%    7.87µs ± 2%     +35.67%  (p=0.001 n=3+3)
FormatEmojiBlock-2      2.21ms ± 2%   13.40ms ± 1%    +507.27%  (p=0.000 n=3+3)
FormatEmojiBlock-4      2.18ms ± 2%   15.24ms ± 3%    +598.44%  (p=0.000 n=3+3)
FormatEmojiBlock-8      2.21ms ± 2%   16.47ms ± 2%    +643.75%  (p=0.000 n=3+3)
FormatEmojiBlock-12     2.32ms ± 3%   18.16ms ± 2%    +683.97%  (p=0.000 n=3+3)

name                  old speed      new speed      delta
FormatEmojiString-2   37.6MB/s ± 2%  27.8MB/s ± 2%     -25.97%  (p=0.000 n=3+3)
FormatEmojiString-4   36.5MB/s ± 2%  25.7MB/s ± 3%     -29.42%  (p=0.000 n=3+3)
FormatEmojiString-8   33.8MB/s ± 2%  25.1MB/s ± 2%     -25.76%  (p=0.000 n=3+3)
FormatEmojiString-12  32.3MB/s ± 1%  23.8MB/s ± 2%     -26.28%  (p=0.000 n=3+3)
FormatEmojiBlock-2    10.1MB/s ± 2%   1.7MB/s ± 1%     -83.55%  (p=0.000 n=3+3)
FormatEmojiBlock-4    10.2MB/s ± 2%   1.5MB/s ± 3%     -85.65%  (p=0.000 n=3+3)
FormatEmojiBlock-8    10.1MB/s ± 2%   1.4MB/s ± 1%     -86.56%  (p=0.000 n=3+3)
FormatEmojiBlock-12   9.65MB/s ± 3%  1.23MB/s ± 2%     -87.26%  (p=0.000 n=3+3)

name                  old alloc/op   new alloc/op   delta
FormatEmojiString-2       881B ± 0%     1458B ± 0%     +65.53%  (p=0.000 n=3+3)
FormatEmojiString-4       882B ± 0%     1464B ± 0%     +65.89%  (p=0.000 n=3+3)
FormatEmojiString-8       885B ± 0%     1467B ± 0%     +65.76%  (p=0.000 n=3+3)
FormatEmojiString-12      887B ± 0%     1472B ± 0%     +65.95%  (p=0.000 n=3+3)
FormatEmojiBlock-2       124kB ± 0%   15270kB ± 0%  +12197.48%  (p=0.000 n=3+3)
FormatEmojiBlock-4       124kB ± 0%   15271kB ± 0%  +12194.37%  (p=0.000 n=3+3)
FormatEmojiBlock-8       124kB ± 0%   15271kB ± 0%  +12191.44%  (p=0.000 n=3+3)
FormatEmojiBlock-12      124kB ± 0%   15272kB ± 0%  +12188.00%  (p=0.000 n=3+3)

name                  old allocs/op  new allocs/op  delta
FormatEmojiString-2       17.0 ± 0%      25.0 ± 0%        ~     (zero variance)
FormatEmojiString-4       17.0 ± 0%      25.0 ± 0%        ~     (zero variance)
FormatEmojiString-8       17.0 ± 0%      25.0 ± 0%        ~     (zero variance)
FormatEmojiString-12      17.0 ± 0%      25.0 ± 0%        ~     (zero variance)
FormatEmojiBlock-2       2.30k ± 0%     3.47k ± 0%        ~     (zero variance)
FormatEmojiBlock-4       2.30k ± 0%     3.47k ± 0%        ~     (zero variance)
FormatEmojiBlock-8       2.30k ± 0%     3.47k ± 0%        ~     (zero variance)
FormatEmojiBlock-12      2.30k ± 0%     3.47k ± 0%        ~     (zero variance)
```

- The original method took about 35% more time than the new PR one, and ~500% more on extreme cases
- The original method took about 65% more allocation operations than the new PR one, and ~1200% more on extreme cases

To verify, 

- check out commit `0cd471d` _"add Benchmarks for FormatEmoji"_
- run `go test -benchmem -run ^$ -bench ^BenchmarkFormatEmoji -cpu 2,4,8,12 -count 3 | tee /tmp/org.log`
- check out latest (commit `e2dfb62` _"simplify FormatEmoji() algorithm"_)
- run `go test -benchmem -run ^$ -bench ^BenchmarkFormatEmoji -cpu 2,4,8,12 -count 3 | tee /tmp/new.log`
- finally run `benchstat -delta-test ttest /tmp/new.log /tmp/org.log`, where `benchstat` is from https://cs.opensource.google/go/x/perf